Re-Enable the ability to Linked Clone from a template
I am not convinced that linked cloning from a template is actually unsupported.  It works perfectly fine on our vSphere 6.0 clusters and I feel that the protection that Templates provide is even more nessesary for Linked Clones as changes to the root snapshots would cause major problems for all linked clones.